requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
cloudinary==1.33.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.5.0
